Recode complete

+++ main/extras/mixins.py
@@ -1,10 +1,10 @@
 import json
 
-from django.http import HttpResponse, HttpResponseForbidden, HttpResponseBadRequest
+from django.http import HttpResponse, HttpResponseForbidden, HttpResponseBadRequest, HttpResponseNotAllowed
 from django.shortcuts import get_object_or_404
 
 
-class RESTfulViewMixin:
+class RESTfulMethods:
 
     @staticmethod
     def build_error_dict(form_errors):
@@ -22,9 +22,22 @@ class RESTfulViewMixin:
 
         return HttpResponse(json.dumps(response), content_type='application/vnd.api+json')
 
+
+class RESTfulViewMixin(RESTfulMethods):
+
+    methods = ['POST', 'GET', 'PATH', 'DELETE']
+
+    def _new_instance(self):
+
+        return getattr(self, 'model')()
+
+    def _get_instance(self, obj_id):
+
+        return get_object_or_404(getattr(self, 'model'), pk=obj_id)
+
     def _save_instance(self, request, instance):
 
-        form = getattr(self, 'form_class')(request.JSON.get('data', {}).get('attributes'), instance=instance)
+        form = getattr(self, 'form')(request.JSON.get('data', {}).get('attributes'), instance=instance)
 
         if form.is_valid():
 
@@ -38,27 +51,19 @@ class RESTfulViewMixin:
 
         return response
 
+    def _get_queryset(self, request, kwargs):
+
+        return getattr(self, 'model').objects.all()
+
     def post(self, request, **kwargs):
 
-        obj = getattr(self, 'model_class')()
+        if 'POST' in self.methods:
 
-        if obj.perms(request.user)['editable']:
+            obj = self._new_instance()
 
-            return self._api_response(self._save_instance(request, obj))
+            if obj.perms(request.user)['editable']:
 
-        else:
-
-            return HttpResponseForbidden()
-
-    def get(self, request, **kwargs):
-
-        if 'obj_id' in kwargs:
-
-            obj = get_object_or_404(getattr(self, 'model_class'), pk=kwargs.get('obj_id'))
-
-            if obj.perms(request.user)['readable']:
-
-                return self._api_response({'data': obj.serialize(request.JSON.get('fields'), request.user)})
+                return self._api_response(self._save_instance(request, obj))
 
             else:
 
@@ -66,65 +71,115 @@ class RESTfulViewMixin:
 
         else:
 
-            data = list()
+            return HttpResponseNotAllowed(self.methods)
 
-            filter_pattern = request.JSON.get('filter')
+    def get(self, request, **kwargs):
 
-            for obj in getattr(self, 'model_class').objects.all():
+        if 'GET' in self.methods:
 
-                if all({not filter_pattern or obj.name.find(filter_pattern) > -1, obj.perms(request.user)['readable']}):
+            if 'obj_id' in kwargs:
 
-                    data.append(obj.serialize(request.JSON.get('fields'), request.user))
+                obj = self._get_instance(kwargs['obj_id'])
 
-            return self._api_response({'data': data})
+                if obj.perms(request.user)['readable']:
 
-    def patch(self, request, **kwargs):
+                    response = {'data': obj.serialize(request.JSON.get('fields'), request.user)}
 
-        obj = get_object_or_404(getattr(self, 'model_class'), pk=kwargs.get('obj_id'))
+                    if hasattr(self, 'included'):
 
-        if obj.perms(request.user)['editable']:
+                        response['included'] = self.included(request, obj)
 
-            return self._api_response(self._save_instance(request, obj))
+                    return self._api_response(response)
+
+                else:
+
+                    return HttpResponseForbidden()
+
+            else:
+
+                if 'draw' in request.GET and hasattr(self, 'datatable_handler'):
+
+                    handler = getattr(self, 'datatable_handler')(request, self._get_queryset(request, kwargs))
+
+                    return self._api_response(handler.build_response())
+
+                else:
+
+                    data = list()
+
+                    filter_pattern = request.JSON.get('filter')
+
+                    for obj in self._get_queryset(request, kwargs):
+
+                        if all({not filter_pattern or obj.name.find(filter_pattern) > -1, obj.perms(request.user)['readable']}):
+
+                            data.append(obj.serialize(request.JSON.get('fields'), request.user))
+
+                    return self._api_response({'data': data})
 
         else:
 
-            return HttpResponseForbidden()
+            return HttpResponseNotAllowed(self.methods)
+
+    def patch(self, request, **kwargs):
+
+        if 'PATCH' in self.methods:
+
+            obj = self._get_instance(kwargs['obj_id'])
+
+            if obj.perms(request.user)['editable']:
+
+                return self._api_response(self._save_instance(request, obj))
+
+            else:
+
+                return HttpResponseForbidden()
+
+        else:
+
+            return HttpResponseNotAllowed(self.methods)
 
     def delete(self, request, **kwargs):
 
-        if 'obj_id' in kwargs:
+        if 'DELETE' in self.methods:
 
-            obj = get_object_or_404(getattr(self, 'model_class'), pk=kwargs.get('obj_id'))
+            if 'obj_id' in kwargs:
 
-            if obj.perms(request.user)['deletable']:
+                obj = self._get_instance(kwargs['obj_id'])
 
-                obj.delete()
+                if obj.perms(request.user)['deletable']:
+
+                    obj.delete()
+
+                    return HttpResponse(status=204)
+
+                else:
+
+                    return HttpResponseForbidden()
+
+            elif 'data' in request.JSON:
+
+                id_list = list()
+
+                for obj_dict in request.JSON.get('data'):
+
+                    obj = self._get_instance(obj_dict['id'])
+
+                    if obj.perms(request.user)['deletable']:
+
+                        id_list.append(obj_dict['id'])
+
+                getattr(self, 'model').objects.filter(pk__in=id_list).delete()
 
                 return HttpResponse(status=204)
 
             else:
 
-                return HttpResponseForbidden()
-
-        elif 'data' in request.JSON:
-
-            id_list = list()
-
-            for obj_dict in request.JSON.get('data'):
-
-                obj = getattr(self, 'model_class').objects.get(pk=obj_dict['id'])
-
-                if obj.perms(request.user)['deletable']:
-
-                    id_list.append(obj_dict['id'])
-
-            getattr(self, 'model_class').objects.filter(pk__in=id_list).delete()
-
-            return HttpResponse(status=204)
+                return HttpResponseBadRequest()
 
         else:
 
-            return HttpResponseBadRequest()
+            return HttpResponseNotAllowed(self.methods)
 
 
 class RESTfulModelMixin:
@@ -134,7 +189,7 @@ class RESTfulModelMixin:
 
         return '/'.join([getattr(self, 'route'), str(getattr(self, 'id'))])
 
-    def _build_filtered_dict(self, fields, **kwargs):
+    def _serialize_data(self, fields, **kwargs):
 
         data = kwargs['data'] if 'data' in kwargs else {'id': getattr(self, 'id'), 'type': getattr(self, 'type')}
